Implement GUI
If it's run without arguments it will show the gui normally. If it is called with one argument it will be considered input json and the schedule will be outputed in the terminal. If called with two arguements, it will save it in the second argument. 

Optimize Interface is not implemented, It needs to be implemented!